requirements.txt
pandas==2.2.2 
numpy==1.26.4 
python-binance==1.0.28 
filelock==3.15.4 
requests==2.32.3 
matplotlib==3.9.2 
pandas_ta==0.3.14b0 
scikit-optimize==0.10.2